// kits/arvel-ecommerce-kit/frontend/src/api/schemas/productCardOut.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */

export interface ProductCardOut {
  id: string
  name: string
  slug: string
  short_description: string | null
  price: number
  stock: number
  original_price: number | null
  thumbnail_url: string | null
  image_srcset: string
  image_sizes: string
  rating: number | null
  rating_count: number | null
  is_new: boolean
  is_bestseller: boolean
  category_id: string | null
  category_name: string | null
  category_slug: string | null
  category_parent_id: string | null
  parent_category_name: string | null
  parent_category_slug: string | null
  vendor_id: string | null
  vendor_name: string | null
  vendor_slug: string | null
}